refactor: split comment thread components
